+++ vehicle/OVMS.X/net.h
@@ -131,7 +131,7 @@ extern unsigned char net_fnbits;               // Net functionality bits
 // (such as if the user explicitely requests a charge to be stopped).
 // The idea is that requesters set these bits, and notifiers clear them.
 extern unsigned int  net_notify_errorcode;     // An error code to be notified
-extern unsigned int  net_notify_errordata;     // Ancilliary data
+extern unsigned long net_notify_errordata;     // Ancilliary data
 extern unsigned int  net_notify_lasterrorcode; // Last error code to be notified
 extern unsigned char net_notify_lastcount;     // A counter used to clear error codes
 extern unsigned int  net_notify;               // Bitmap of notifications outstanding
@@ -178,7 +178,7 @@ void net_state_enter(unsigned char);
 void net_state_activity(void);
 void net_state_ticker(void);
 
-void net_req_notification_error(unsigned int errorcode, unsigned int errordata);
+void net_req_notification_error(unsigned int errorcode, unsigned long errordata);
 void net_req_notification(unsigned int notify);
 
 #endif // #ifndef __OVMS_NET_H
